Moved the agent-env association performance test utility from production code to test-shared.
