+++ server/src/test-shared/java/com/thoughtworks/go/server/perf/commands/RegisterAgentCommand.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.thoughtworks.go.server.service.perf.commands;
+package com.thoughtworks.go.server.perf.commands;
 
 import com.thoughtworks.go.config.Agent;
 import com.thoughtworks.go.server.service.AgentRuntimeInfo;
